+++ app/Http/Controllers/V2/EnrollmentController.php
@@ -6,6 +6,7 @@ use App\Http\Controllers\Controller;
 use App\Interfaces\EnrollmentRepositoryInterface;
 use Illuminate\Http\Request;
 use App\Http\Requests\UpdateEnrollmentRequest;
+use Illuminate\Support\Facades\Auth;
 
 class EnrollmentController extends Controller
 {
@@ -21,6 +22,13 @@ class EnrollmentController extends Controller
     {
         try {
             $user = $request->user();
+
+            $existingEnrollment = $this->enrollmentRepository->getEnrollmentByUserAndCourse($user->id, $courseId);
+            if ($existingEnrollment) {
+                return response()->json(['message' => 'User is already enrolled in this course.'], 400);
+            }
+
+            // Enregistrer l'inscription
             $enrollment = $this->enrollmentRepository->enrollUser($user->id, $courseId);
             return response()->json(['message' => 'Enrollment successful.', 'data' => $enrollment], 201);
         } catch (\Exception $e) {
@@ -32,6 +40,10 @@ class EnrollmentController extends Controller
     public function listEnrollments($courseId)
     {
         try {
+            if (!Auth::user()->hasRole(['mentor', 'admin'])) {
+                return response()->json(['message' => 'Unauthorized'], 403);
+            }
+
             $enrollments = $this->enrollmentRepository->getEnrollmentsByCourse($courseId);
             return response()->json(['data' => $enrollments]);
         } catch (\Exception $e) {
@@ -43,6 +55,10 @@ class EnrollmentController extends Controller
     public function updateEnrollment(UpdateEnrollmentRequest $request, $enrollmentId)
     {
         try {
+            if (!Auth::user()->hasRole(['mentor', 'admin'])) {
+                return response()->json(['message' => 'Unauthorized'], 403);
+            }
+
             $enrollment = $this->enrollmentRepository->updateEnrollmentStatus($enrollmentId, $request->status);
             return response()->json(['message' => 'Enrollment updated.', 'data' => $enrollment]);
         } catch (\Exception $e) {
@@ -54,6 +70,10 @@ class EnrollmentController extends Controller
     public function deleteEnrollment(Request $request, $enrollmentId)
     {
         try {
+            if (!Auth::user()->hasRole(['mentor', 'admin'])) {
+                return response()->json(['message' => 'Unauthorized'], 403);
+            }
+
             $this->enrollmentRepository->deleteEnrollment($enrollmentId);
             return response()->json(['message' => 'Enrollment deleted.'], 204);
         } catch (\Exception $e) {
